third_party/blink/web_tests/external/wpt/css/css-ui/parsing/resize-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['resize'] = "auto" should not set the property value
  assert_equals: expected "" but got "auto"
Harness: the test ran to completion.
